250822 박건아 중고거래 프론트컨트롤러 작업중

+++ SafeDugout_backend/src/main/java/com/bullPenTalk/app/dto/BatterRecordDTO.java
@@ -140,7 +140,7 @@ public class BatterRecordDTO {
 	public void setSRBI(int sRBI) {
 		SRBI = sRBI;
 	}
-	public int getOPS() {
+	public double getOPS() {
 		return OPS;
 	}
 	public void setOPS(int oPS) {
